Create endpoint for teams.
Ran into a few issues regarding the difference between string keys and atom keys.

Alters create team to take an id for the owner and retrieve it from the database. Realize
now that it doesn't make sense to send the whole user object with the request ¯\_(ツ)_/¯.

Avoided writing integration tests because I want to move quickly to get to the interesting parts
of this project (the channels part).