Update all deps and fix subsequent broken code
The :on_replace option for `belongs_to :owner` on the Team schema
needed to be set to :nilify, which was not apparent until a change
to ecto was made that would raise if you passed put_assoc a struct
if you have :on_replace set to :update.

:nilify will set the FK (owner_id) to nil, and then set it to the new
value. This is not very clear.