Room endpoints
Modifies some of the behavior from the context generator, not totally
sure which way is proper but this resembles my other code somewhat more.

Note, these controller tests along with some of the other recent ones
are only testing the happy path status codes and not the actual response
bodies.