Return user along with the token on authentication and self endpoint
It made more sense to return the user along with the token, and this
surfaced the issue that the token was not really being serialized
correctly in the first place. In order to be able to load the user
there needed to be a token struct made.

Created a self endpoint that would return the current user, which
is the user for which the auth token sent with the request corresponds.