Introduce elm to the project
react-elm-components was very out of date, so we needed to fork and
it in order to get it to work. This was due to React.createClass being
deprecated in v16.x.x. This was a problem because react-elm-components
had react as a peer-dependency and not a dependency.

Currently there is a copy/paste elm component, but that will change soon.